requirements.txt
pycryptodome
pycryptodome>=3.20.0